fix!: Windows locate_executable finds wrong binary to run
This fixes issue #140, by calling a Python subprocess to
run `shutil.where` instead of a cmd subprocess to run `where`.

It is a breaking change, because it introduces a new
requirement that when impersonating a user for subprocesses, the
Python installation hosting the library can be run by the
impersonated user as well.

Signed-off-by: Mark Wiebe <399551+mwiebe@users.noreply.github.com>