fix: ensure process exit code is a 32-bit signed integer (#148)
Problem:
We've observed on Windows systems that a process can return an exit code
that fits in a 32-bit integer, but it is interpreted as an unsigned
integer rather than a signed integer. e.g. 0xFFFD0000 as 4294770688
rather than -199608.

Solution:
Clamp the return code to 32-bit and ensure that the bits are interpreted
as a signed integer.

Signed-off-by: Daniel Neilson <53624638+ddneilson@users.noreply.github.com>